Backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
psycopg2-binary==2.9.5
PyJWT==2.6.0
bcrypt==4.0.1
pygeodesic==0.1.8
gunicorn==20.1.0
geopy==2.3.0
